DDF-4436 fixed queries to use default search forms
