+++ ui/packages/catalog-ui-search/src/main/webapp/component/query-advanced/query-advanced.view.js
@@ -23,6 +23,7 @@ const FilterBuilderModel = require('../filter-builder/filter-builder.js')
 const cql = require('../../js/cql.js')
 const store = require('../../js/store.js')
 const QuerySettingsView = require('../query-settings/query-settings.view.js')
+const user = require('../singletons/user-instance.js')
 
 module.exports = Marionette.LayoutView.extend({
   template: template,
@@ -41,6 +42,51 @@ module.exports = Marionette.LayoutView.extend({
   initialize: function() {
     this.$el.toggleClass('is-form-builder', this.options.isFormBuilder === true)
     this.$el.toggleClass('is-form', this.options.isForm === true)
+    this.listenTo(
+      user.getQuerySettings(),
+      'change:template',
+      this.handleTemplateChange
+    )
+    this.handleTemplateChange()
+  },
+  handleTemplateChange: function() {
+    let userDefaultTemplate = user.getQuerySettings().get('template')
+    if (userDefaultTemplate) {
+      let sorts =
+        userDefaultTemplate['querySettings'] &&
+        userDefaultTemplate['querySettings'].sorts
+      if (sorts) {
+        sorts = sorts.map(sort => ({
+          attribute: sort.split(',')[0],
+          direction: sort.split(',')[1],
+        }))
+      }
+      this.model.set({
+        type: 'custom',
+        title: userDefaultTemplate['title'],
+        filterTree: userDefaultTemplate['filterTemplate'],
+        src:
+          (userDefaultTemplate['querySettings'] &&
+            userDefaultTemplate['querySettings'].src) ||
+          '',
+        federation:
+          (userDefaultTemplate['querySettings'] &&
+            userDefaultTemplate['querySettings'].federation) ||
+          'enterprise',
+        sorts: sorts,
+        'detail-level':
+          (userDefaultTemplate['querySettings'] &&
+            userDefaultTemplate['querySettings']['detail-level']) ||
+          'allFields',
+      })
+      if (
+        this.options.isForm === true &&
+        this.model.get('filterTree') !== undefined &&
+        this.queryAdvanced.currentView !== undefined
+      ) {
+        this.queryAdvanced.currentView.deserialize(this.model.get('filterTree'))
+      }
+    }
   },
   onBeforeShow: function() {
     this.model = this.model._cloneOf
